replace kdam progress bar with ConEmu OSC (only when isatty)
The kdam bar stuff I hacked together was really quite terrible
(especially the work I did to let the logger continue printing without
breaking the bar). Also, because I didn't check isatty on stdout, it was
filling up the systemd logs with garbage.

This is not an objective improvement, because we lose 1) the actual
numbers; 2) the speed at which we are processing works; but that's a
tradeoff I'm willing to make.

Note that this also does not work in all terminals, and I haven't
bothered to check whether the terminal supports it first (there's no
easy way to do so, you'd need to check `TERM` i think?). It works
properly in Ghostty. Alacritty properly parses the codes but does
nothing with them. Some other terminals (iTerm, WezTerm I think)
interpret these as iTerm notification OSCs and will spam the user with
many many notifications with useless content.